+++ src/UI/TransactionInfo/components/DataField/components/index.tsx
@@ -1,2 +1,2 @@
-export * from './Anchorme';
+export * from './Linkified';
 export * from './ModalLink';