+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[2.2.1](https://github.com/ElrondNetwork/dapp-core/pull/510)] - 2022-12-16
+- [Changed logging in logic to allow signing custom token](https://github.com/ElrondNetwork/dapp-core/pull/509)
+- [Security issue fixed by substituting `anchorme` with linkify.js and fixed `getScamFlag` logic](https://github.com/ElrondNetwork/dapp-core/pull/508)
+
 ## [[2.2.0](https://github.com/ElrondNetwork/dapp-core/pull/507)] - 2022-12-14
 - [Added Axios interceptor wrapper to be used with `nativeAuth`](https://github.com/ElrondNetwork/dapp-core/pull/506)
 - [Changed `signTransactions.ts` import to resolve Next.js support](https://github.com/ElrondNetwork/dapp-core/pull/505)